# api_spec.yaml
openapi: "3.0.0"
info:
  title: "APIs ServiEstado"
  version: "1.0.0"
  description: "Documentación relacionada a las APIs utilizadas para el nuevo Portal Personas de ServiEstado."
servers:
  - url: https://desa-plataformadigital.banco.bestado.cl/bff/v1/se-personas-bff
  - url: http://localhost:6969
  - url: https://apis.serviestado.cl/v1
...
paths:
...
  /convenios/{id}:
    get:
      tags:
      - "Convenios"
      summary: "Obtiene el detalle de un convenio ServiEstado"
      description: ""
      operationId: "findConveniosById"
      parameters:
        - name: id
          in: path
          description: ID del convenio
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Ok
          content:
            application/json:    
              schema:
                $ref: "#/components/schemas/DetalleConvenio"
...
components:
  schemas:
    Rubro:
      type: "object"
      properties:
        idRubro:
          type: "integer"
          example: 10000
        nombreRubro:
          type: "string"
          example: "Telefonía Hogar"
        logotipo:
          type: "string"
          example: "https://ww2.movistar.cl/inc_header_footer_v8/assets/img/logo-movistar.svg"
    Convenio:
      type: "object"
      properties:
        idConvenio:
          type: "integer"
          example: 101
        servicio:
          type: "string"
          example: "Movistar Hogar"
        empresa:
          type: "string"
          example: "Movistar"
        logotipo:
          type: "string"
          example: "https://ww2.movistar.cl/inc_header_footer_v8/assets/img/logo-movistar.svg"
        rubro:
          $ref: "#/components/schemas/Rubro"
    AyudaConvenio:
      type: "object"
      properties:
        nombreIdentificador:
          type: "string"
          example: "N° de cliente"
        tipoCodigo:
          type: "string"
          example: "QR"
        imagenAyuda:
          type: "string"
          example: "https://ww2.movistar.cl/inc_header_footer_v8/assets/img/logo-movistar.svg"
        busqueda:
            type: "array"
            items:
              $ref: "#/components/schemas/busquedaConvenio"
    busquedaConvenio:
      type: "object"
      properties:
        nombre:
          type: "string"
          example: "telefono"
        tipo:
          type: "string"
          example: "numerico"
...
    DetalleConvenio:
      allOf:
          - $ref: "#/components/schemas/Convenio"
          - type: "object"
            properties:
              ayuda:
                $ref: "#/components/schemas/AyudaConvenio"
...